Add files via upload
The initial step I implemented during the clean up of Amazon Sales Data 2025 data frame was to drop the unobserved values and duplicate rows as the first step in preparing the data for analysis Once I had cleaned the data entries, I standardized the column headers by changing all headers to lowercase and replacing spaces with underscores—for example, I converted Order Id to order_id. This led to better consistency and made it easier to work with the dataset programmatically.

This was followed by the date column, which was initially in text format. I used Excel’s “Text to Columns” tool to convert this into a consistent dd-mm-yyyy date format, making it much easier to sort and filter based on dates.

Then I cleaned text values within categorical columns to make them more readable. The payment_method column encoded information in the form of short codes like D and C which I translated to more descriptive labels like Debit Card and Credit Card. Similarly, in the status column, Cd and P were expanded as Cash on Delivery and Paid.

Last but not least, I made sure that all of the numeric columns — price, quantity and total_sales — were formatted as numbers and the categorical fields were defined as text. Together, these steps cleaned and structured the dataset in a way that made it more fit for analysis or visualization.